+++ src/Make_dice.mak
@@ -33,6 +33,7 @@ SRC = \
 	charset.c \
 	crypt.c \
 	crypt_zip.c \
+	debugger.c \
 	dict.c \
 	diff.c \
 	digraph.c \
@@ -93,6 +94,7 @@ OBJ =	o/arabic.o \
 	o/charset.o \
 	o/crypt.o \
 	o/crypt_zip.o \
+	o/debugger.o \
 	o/dict.o \
 	o/diff.o \
 	o/digraph.o \
@@ -178,6 +180,8 @@ o/charset.o:	charset.c  $(SYMS)
 o/crypt.o:	crypt.c  $(SYMS)
 
 o/crypt_zip.o:	crypt_zip.c  $(SYMS)
+
+o/debugger.o:	debugger.c  $(SYMS)
 
 o/dict.o:	dict.c	$(SYMS)
