+++ src/Make_ivc.mak
@@ -217,6 +217,7 @@ LINK32_OBJS= \
 	"$(INTDIR)/charset.obj" \
 	"$(INTDIR)/crypt.obj" \
 	"$(INTDIR)/crypt_zip.obj" \
+	"$(INTDIR)/debugger.obj" \
 	"$(INTDIR)/dict.obj" \
 	"$(INTDIR)/diff.obj" \
 	"$(INTDIR)/digraph.obj" \
@@ -366,6 +367,10 @@ SOURCE=.\crypt.c
 # Begin Source File
 
 SOURCE=.\crypt_zip.c
+# End Source File
+# Begin Source File
+
+SOURCE=.\debugger.c
 # End Source File
 # Begin Source File
