+++ src/proto.h
@@ -68,6 +68,7 @@ extern int _stricoll(char *a, char *b);
 # ifdef FEAT_CSCOPE
 #  include "if_cscope.pro"
 # endif
+# include "debugger.pro"
 # include "dict.pro"
 # include "diff.pro"
 # include "digraph.pro"
